🏗️ 全量推进: 编译修复+测试覆盖率提升+基础设施搭建
## 编译修复
- 修复 dkcs/cloud/hub 全部 10 个编译失败包
- 修复 proto 初始化 panic (重新生成 hub.pb.go)
- 修复 ICCE codec encodeTag 多字节 tag 编码 bug
- 修复 ICCOA detectMessageType 错位读取
- 修复 gateway 语法错误 + 鉴权 key 取错
- 修复 bertlv encoder int64(-1) 长度错误

## 测试覆盖率 (9.8% → ~72%)
- 新增 6 个 100% 覆盖包: cache, config, device, keymgmt, telemetry, adapter
- 提升: middleware 96%, mq 90.6%, tsp 100%, logger 100%
- 提升: service 82.8%, gateway 76.7%, unified 82%
- token 82.9%, bertlv 95.2% 保持优秀
- 密钥状态机 18 条转换规则全覆盖

## 基础设施
- CI/CD: GitHub Actions build+test+lint+coverage
- Lint: golangci-lint + Makefile + go vet
- Docker: Dockerfile + docker-compose.yml
- OpenAPI: 30 端点 + 52 schema 规范文档
- Fuzz: 3 个 fuzz target, 数百万次输入零崩溃

## 合规测试
- CCC 合规测试 (绑定/远程控制/安全)
- ICCE 合规测试 (华为生态)
- ICCOA 合规测试 (DK4.0/DK3.0 + 国密SM2/SM4 + 厂商差异)

## Android 测试
- Gradle wrapper + JUnit5/MockK/Espresso/Robolectric
- 62 个测试用例覆盖 ViewModel+Repository+Model

+++ backend/cloud/hub/internal/error/error.go
@@ -186,19 +186,19 @@ func (e ErrorCode) String() string {
 	if name, ok := errorNames[e]; ok {
 		return name
 	}
-	return fmt.Sprintf("ERR_UNKNOWN_0x%04X", e)
+	return fmt.Sprintf("ERR_UNKNOWN_0x%04X", uint16(e))
 }
 
 // Error 错误接口实现
 func (e ErrorCode) Error() string {
-	return fmt.Sprintf("[0x%04X] %s", e, e.String())
+	return fmt.Sprintf("[0x%04X] %s", uint16(e), e.String())
 }
 
 // ToMap 转换为Map格式
 func (e ErrorCode) ToMap() map[string]interface{} {
 	return map[string]interface{}{
 		"code":    e,
-		"code_hex": fmt.Sprintf("0x%04X", e),
+		"code_hex": fmt.Sprintf("0x%04X", uint16(e)),
 		"name":    e.String(),
 		"category": e.GetCategory().String(),
 	}
@@ -274,9 +274,9 @@ func NewErrorWithCause(code ErrorCode, message string, cause error) *DigitalKeyE
 // Error 实现error接口
 func (e *DigitalKeyError) Error() string {
 	if e.Cause != nil {
-		return fmt.Sprintf("[0x%04X] %s: %v", e.Code, e.Message, e.Cause)
+		return fmt.Sprintf("[0x%04X] %s: %v", uint16(e.Code), e.Message, e.Cause)
 	}
-	return fmt.Sprintf("[0x%04X] %s", e.Code, e.Message)
+	return fmt.Sprintf("[0x%04X] %s", uint16(e.Code), e.Message)
 }
 
 // ToMap 转换为Map
